+++ R/mr_radialegger_stan.R
@@ -1,0 +1,76 @@
+#' Bayesian inverse variance weighted model with a choice of prior distributions fitted using Stan.
+#'
+#' Bayesian inverse variance weighted model with a choice of prior distributions fitted using Stan
+#'
+#' @param data A data of class [`mr_format`].
+#' @param prior An integer for selecting the prior distributions;
+#'
+#' * `1` selects a non-informative set of priors;
+#' * `2` selects weakly informative priors;
+#' * `3` selects a pseudo-horseshoe prior on the causal effect;
+#' * `4` selects joint prior of the intercept and causal effect estimate.
+#' @param n.chains Numeric indicating the number of chains used in the HMC estimation in rstan, the default is `3` chains.
+#' @param n.burn Numeric indicating the burn-in period of the Bayesian HMC estimation. The default is `1000` samples.
+#' @param n.iter Numeric indicating the number of iterations in the Bayesian HMC estimation. The default is `5000` iterations.
+#' @param rho Numeric indicating the correlation coefficient input into the joint prior distribution. The default is `0.5`.
+#' @param seed Numeric indicating the random number seed. The default is `12345`.
+#' @param ... Additional arguments passed through to [`rstan::sampling()`].
+#'
+#' @return An object of class [`stanfit`].
+#'
+#' @export
+#' @references Bowden, J., et al., Improving the visualization, interpretation and analysis of two-sample summary data Mendelian randomization via the Radial plot and Radial regression. International Journal of Epidemiology, 2018. 47(4): p. 1264-1278. <https://doi.org/10.1093/ije/dyy101>.
+#'
+#' @examples
+#' radegger_fit <- mr_radialegger_stan(bmi_insulin)
+#' print(radegger_fit)
+mr_radialegger_stan <- function(data,
+                          prior = 1,
+                          n.chains = 3,
+                          n.burn = 1000,
+                          n.iter = 5000,
+                          rho = 0.5,
+                          seed = 12345,
+                          ...) {
+
+  # check class of object
+  if (!("mr_format" %in% class(data))) {
+    stop(
+      'The class of the data object must be "mr_format", please resave the object with the output of e.g. object <- mr_format(object).'
+    )
+  }
+
+  pars <- c("intercept","estimate","sigma")
+
+  # converting dataset to a list
+  datam <- list(
+    n = nrow(data),
+    xbeta = data[, 2] / data[, 5],
+    ybeta = data[, 3] / data[, 5],
+    prior = prior,
+    rho = rho
+  )
+
+  radialeggerfit <- rstan::sampling(
+    object = stanmodels$mrradialegger,
+    data = datam,
+    pars = pars,
+    chains = n.chains,
+    warmup = n.burn,
+    iter = n.iter,
+    seed = seed,
+    control = list(adapt_delta = 0.999, max_treedepth = 15),
+    ...
+  )
+
+  return(radialeggerfit)
+
+}
+
+
+data(bmi_insulin)
+dat <- mr_format(rsid = bmi_insulin[,"rsid"],
+           xbeta = bmi_insulin[,"beta.exposure"],
+           ybeta = bmi_insulin[,"beta.outcome"],
+           xse = bmi_insulin[,"se.exposure"],
+           yse = bmi_insulin[,"se.outcome"])