+++ NEWS.md
@@ -1,6 +1,8 @@
 # mrbayes 0.2.0
 
 * IVW implemented using rstan.
+* MR-Egger implemented using rstan.
+* Radial MR-Egger implemented using rstan.
 
 # mrbayes 0.1.0